requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
